+++ src/shared/ui/Modal/Modal.tsx
@@ -1,36 +1,18 @@
-import { useCallback } from "react";
-import type { FC, ReactNode } from "react";
-import styles from "./Modal.module.scss";
+import type { FC } from "react";
 import { createPortal } from "react-dom";
+import Portal from "@/shared/ui/Portal/Portal";
 
 type ModalProps = {
 	isOpen: boolean;
 	onClose: () => void;
-	children: ReactNode;
 };
 
-const Modal: FC<ModalProps> = ({ isOpen, onClose, children }) => {
-	const handleCloseButtonClick = useCallback(() => {
-		onClose();
-	}, [onClose]);
-
+const Modal: FC<ModalProps> = ({ isOpen, onClose }) => {
 	let rootEl = document.querySelector("#root");
 	if (!rootEl) rootEl = document.body;
 
 	if (!isOpen) return null;
-	return createPortal(
-		<div className={styles.modal}>
-			<div className={styles.modal__content}>
-				<button className={styles.modal__close} onClick={handleCloseButtonClick}>
-					<svg xmlns="http://www.w3.org/2000/svg" x="0" y="0" viewBox="0 0 50 50">
-						<path d="M 7.71875 6.28125 L 6.28125 7.71875 L 23.5625 25 L 6.28125 42.28125 L 7.71875 43.71875 L 25 26.4375 L 42.28125 43.71875 L 43.71875 42.28125 L 26.4375 25 L 43.71875 7.71875 L 42.28125 6.28125 L 25 23.5625 Z" />
-					</svg>
-				</button>
-				{children}
-			</div>
-		</div>,
-		rootEl
-	);
+	return createPortal(<Portal onClose={onClose} />, rootEl);
 };
 
 export default Modal;
