--- src/shared/ui/Portal/Modal.module.scss
@@ -1,29 +1,0 @@
-.modal {
-	display: flex;
-	position: fixed;
-	top: 0;
-	left: 0;
-	justify-content: center;
-	align-items: center;
-	width: 100%;
-	height: 100%;
-
-	&__content {
-		position: relative;
-		max-width: 900px;
-		padding: 30px;
-		margin: 0 30px;
-		background: white;
-    border: 3px solid black;
-	}
-
-	&__close {
-		position: absolute;
-		top: 10px;
-		right: 10px;
-		width: 20px;
-		background-color: transparent;
-		border: none;
-		cursor: pointer;
-	}
-}
