+++ src/shared/ui/Portal/Portal.module.scss
@@ -14,7 +14,7 @@
 		padding: 30px;
 		margin: 0 30px;
 		background: white;
-    border: 3px solid black;
+		border: 3px solid black;
 	}
 
 	&__close {
@@ -25,5 +25,15 @@
 		background-color: transparent;
 		border: none;
 		cursor: pointer;
+	}
+
+	&__title {
+		font-size: 24px;
+	}
+
+	&__text {
+		margin-top: 10px;
+		font-size: 16px;
+		text-align: justify;
 	}
 }
