+++ src/widgets/LayoutHeader/Header.module.scss
@@ -42,16 +42,3 @@
 	font-style: italic;
 	color: red;
 }
-
-
-.modal {
-	&__title {
-		font-size: 24px;
-	}
-
-	&__text {
-		margin-top: 10px;
-		font-size: 16px;
-		text-align: justify;
-	}
-}
